hgfhgcfh

+++ DelpinUI/Debtor.cs
@@ -29,18 +29,33 @@ namespace DelpinUI
                 
         //}
         //update table
-        private void ClearTextBox()
+        //private void ClearTextBox()
+        //{
+        //    cprText.Clear();
+        //    cvrText.Clear();
+        //    BnameText.Clear();
+        //    PfnameText.Clear();
+        //    PlnameText.Clear();
+        //    adressText.Clear();
+        //    city.Clear();
+        //    postalcodeText.Clear();
+        //    phoneText.Clear();
+        //    emailText.Clear();
+        //}
+        private void ClearAllTextBoxes()
         {
-            cprText.Clear();
-            cvrText.Clear();
-            BnameText.Clear();
-            PfnameText.Clear();
-            PlnameText.Clear();
-            adressText.Clear();
-            city.Clear();
-            postalcodeText.Clear();
-            phoneText.Clear();
-            emailText.Clear();
+           
+
+            foreach (Control c in this.Controls)
+            {
+                string getType = c.GetType().ToString();
+                if (c.GetType().ToString() == "System.Windows.Forms.TextBox")
+                {
+                    c.Text = "";
+                }
+            }
+
+          
         }
 
 
@@ -98,7 +113,7 @@ namespace DelpinUI
              
 
             }
-            ClearTextBox();
+            ClearAllTextBoxes();
          
 
 
@@ -131,6 +146,50 @@ namespace DelpinUI
 
         }
 
+        private void getBusinessDebtor()
+        {
+            { 
+
+                string cvr = ViewDeb.SelectedRows[0].Cells[0].Value + string.Empty;
+                string bName = ViewDeb.SelectedRows[0].Cells[1].Value + string.Empty;
+                string adresse = ViewDeb.SelectedRows[0].Cells[2].Value + string.Empty;
+                string City = ViewDeb.SelectedRows[0].Cells[3].Value + string.Empty;
+                string pCode = ViewDeb.SelectedRows[0].Cells[4].Value + string.Empty;
+                string phoneN = ViewDeb.SelectedRows[0].Cells[5].Value + string.Empty;
+                string eMail = ViewDeb.SelectedRows[0].Cells[6].Value + string.Empty;
+
+                cvrText.Text = cvr;
+                BnameText.Text = bName;
+                adressText.Text = adresse;
+                city.Text = City;
+                postalcodeText.Text = pCode;
+                phoneText.Text = phoneN;
+                emailText.Text = eMail;
+            }
+        }
+        private void getPersonalDebtor()
+        {
+            {
+
+                string cpr = ViewDeb.SelectedRows[0].Cells[0].Value + string.Empty;
+                string Fname = ViewDeb.SelectedRows[0].Cells[1].Value + string.Empty;
+                string Lname = ViewDeb.SelectedRows[0].Cells[2].Value + string.Empty;
+                string adresse = ViewDeb.SelectedRows[0].Cells[3].Value + string.Empty;
+                string City = ViewDeb.SelectedRows[0].Cells[4].Value + string.Empty;
+                string pCode = ViewDeb.SelectedRows[0].Cells[5].Value + string.Empty;
+                string phoneN = ViewDeb.SelectedRows[0].Cells[6].Value + string.Empty;
+                string eMail = ViewDeb.SelectedRows[0].Cells[7].Value + string.Empty;
+
+                cprText.Text = cpr;
+                PfnameText.Text = Fname;
+                PlnameText.Text = Lname;
+                adressText.Text = adresse;
+                city.Text = City;
+                postalcodeText.Text = pCode;
+                phoneText.Text = phoneN;
+                emailText.Text = eMail;
+            }
+        }
         private void textBox8_TextChanged(object sender, EventArgs e)
         {
 
@@ -151,7 +210,7 @@ namespace DelpinUI
         {
 
         }
-
+        //create debtor and refreshes datagridview
         private void CreateBdeb_Click(object sender, EventArgs e)
         {
             if (radioButton1.Checked)
@@ -199,6 +258,7 @@ namespace DelpinUI
                
             }
             updateDatagridView();
+            ClearAllTextBoxes();
 
         }
 
@@ -224,7 +284,7 @@ namespace DelpinUI
         }
 
         
-
+        // get specific debtor Business or Personal
         private void button1_Click(object sender, EventArgs e)
         {
             if (radioButton1.Checked)
@@ -259,92 +319,169 @@ namespace DelpinUI
 
 
         }
-
+        
         private void UpdateDebtor_Click(object sender, EventArgs e)
         {
-            if(radioButton1.Checked)
+            if (radioButton1.Checked)
             {
-                CreateDeb.Visible = false;
-                SaveDebtor.Visible = true;
-                CancelUpdate.Visible = true;
-                UpdateDebtor.Visible = false;
+                if (Utility.CheckForValidEmail(emailText.Text) == false)
+                {
+                    MessageBox.Show("Ugyldigt E-mailadresse");
+                    return;
+                }
+                const string message = "Vil du rette kunden?";
+                const string caption = "Annuller";
+                var result = MessageBox.Show(message, caption,
+                                             MessageBoxButtons.YesNo,
+                                             MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
+                {
+                    string succes = controller.UpdateBusinessDebtor(cvrText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
+                   phoneText.Text, emailText.Text, cvrText.Text, BnameText.Text);
+                    MessageBox.Show(succes);
+                }
+                else
+                {
+
+                }
+
+
+
+             
+
+
+
+
             }
             if (radioButton2.Checked)
             {
-                CreateDeb.Visible = false;
-                SaveDebtor.Visible = true;
-                CancelUpdate.Visible = true;
-                UpdateDebtor.Visible = false;
+                if (Utility.CheckForValidEmail(emailText.Text) == false)
+                {
+                    MessageBox.Show("Ugyldigt E-mailadresse");
+                    return;
+                }
+                if (Utility.CheckForValidEmail(emailText.Text) == false)
+                {
+                    MessageBox.Show("Ugyldigt E-mailadresse");
+                    return;
+                }
+                const string message = "Vil du rette kunden?";
+                const string caption = "Annuller";
+                var result = MessageBox.Show(message, caption,
+                                             MessageBoxButtons.YesNo,
+                                             MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
+                {
+                    string succes = controller.UpdatePersonalDebtor(cprText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
+                     phoneText.Text, emailText.Text, cprText.Text, PfnameText.Text, PlnameText.Text);
+                    MessageBox.Show(succes);
+
+                }
+                else
+                {
+
+                }
+
+
+
+                
+
+
+
+
+
             }
+            
+            CreateDeb.Visible = true;
+            UpdateDebtor.Visible = true;
+            updateDatagridView();
+            ClearAllTextBoxes();
+            //if(radioButton1.Checked)
+            //{
+            //    CreateDeb.Visible = false;
+            //    SaveDebtor.Visible = true;
+            //    CancelUpdate.Visible = true;
+            //    UpdateDebtor.Visible = false;
+            //}
+            //if (radioButton2.Checked)
+            //{
+            //    CreateDeb.Visible = false;
+            //    SaveDebtor.Visible = true;
+            //    CancelUpdate.Visible = true;
+            //    UpdateDebtor.Visible = false;
+            //}
 
         }
        
 
         private void SaveDebtor_Click(object sender, EventArgs e)
         {
-            if(radioButton1.Checked)
-            {
-                if (Utility.CheckForValidEmail(emailText.Text) == false)
-                {
-                    MessageBox.Show("Ugyldigt E-mailadresse");
-                    return;
-                }
+            //if(radioButton1.Checked)
+            //{
+            //    if (Utility.CheckForValidEmail(emailText.Text) == false)
+            //    {
+            //        MessageBox.Show("Ugyldigt E-mailadresse");
+            //        return;
+            //    }
 
 
-                string succes = controller.UpdateBusinessDebtor(cvrText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
-                    phoneText.Text, emailText.Text, cvrText.Text, BnameText.Text);
-                    MessageBox.Show(succes);
+            //    string succes = controller.UpdateBusinessDebtor(cvrText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
+            //        phoneText.Text, emailText.Text, cvrText.Text, BnameText.Text);
+            //        MessageBox.Show(succes);
                
 
 
                 
-            }
-            if (radioButton2.Checked)
-            {
-                if (Utility.CheckForValidEmail(emailText.Text) == false)
-                {
-                    MessageBox.Show("Ugyldigt E-mailadresse");
-                    return;
-                }
+            //}
+            //if (radioButton2.Checked)
+            //{
+            //    if (Utility.CheckForValidEmail(emailText.Text) == false)
+            //    {
+            //        MessageBox.Show("Ugyldigt E-mailadresse");
+            //        return;
+            //    }
 
 
 
-                string succes = controller.UpdatePersonalDebtor(cprText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
-                    phoneText.Text, emailText.Text, cprText.Text, PfnameText.Text, PlnameText.Text);
-                    MessageBox.Show(succes);
+            //    string succes = controller.UpdatePersonalDebtor(cprText.Text, adressText.Text, Convert.ToInt32(postalcodeText.Text), city.Text,
+            //        phoneText.Text, emailText.Text, cprText.Text, PfnameText.Text, PlnameText.Text);
+            //        MessageBox.Show(succes);
                 
 
 
 
                 
 
-            }
-            CreateDeb.Visible = true;
-            UpdateDebtor.Visible = true;
-            SaveDebtor.Visible = false;
-            CancelUpdate.Visible = false;
-            ClearTextBox();
+            //}
+            //updateDatagridView();
+            //CreateDeb.Visible = true;
+            //UpdateDebtor.Visible = true;
+            //SaveDebtor.Visible = false;
+            //CancelUpdate.Visible = false;
+            //ClearTextBox();
          
         }
-
+        //cancel update function
         private void CancelUpdate_Click(object sender, EventArgs e)
         {
-            CreateDeb.Visible = true;
-            UpdateDebtor.Visible = true;
-            SaveDebtor.Visible = false;
-            CancelUpdate.Visible = false;
-            ClearTextBox();
+            //CreateDeb.Visible = true;
+            //UpdateDebtor.Visible = true;
+            //SaveDebtor.Visible = false;
+            //CancelUpdate.Visible = false;
+            //ClearTextBox();
          
 
 
         }
-
+        //Delete debtor and updates datagridview
         private void DeleteDebtor_Click(object sender, EventArgs e)
         {
             //indsæt slet debtor metode
             if (radioButton1.Checked)
             {
-                //CancelDelete();
+                //CancelBusinessDebtorDelete();
                 const string message = "Er du sikker på du vil slette Kunden?";
                 const string caption = "Annuller";
                 var result = MessageBox.Show(message, caption,
@@ -355,7 +492,7 @@ namespace DelpinUI
                 {
                     string Success = controller.DeleteBusinessDebter(cvrText.Text);
                     MessageBox.Show(Success);
-                    ClearTextBox();
+                    ClearAllTextBoxes();
                 }
                 else
                 {
@@ -366,9 +503,50 @@ namespace DelpinUI
             }
             if (radioButton2.Checked)
             {
-                string Success = controller.DeletePersonalDebtor(cprText.Text);
-                MessageBox.Show(Success);
-                ClearTextBox();
+                //CancelBusinessDebtorDelete();
+                const string message = "Er du sikker på du vil slette Kunden?";
+                const string caption = "Annuller";
+                var result = MessageBox.Show(message, caption,
+                                             MessageBoxButtons.YesNo,
+                                             MessageBoxIcon.Question);
+
+                if (result == DialogResult.Yes)
+                {
+                    string Success = controller.DeletePersonalDebtor(cprText.Text);
+                    MessageBox.Show(Success);
+                    ClearAllTextBoxes();
+                }
+                else
+                {
+
+                }
+               
+            }
+            updateDatagridView();
+        }
+
+        private void ViewDeb_CellMouseClick(object sender, DataGridViewCellMouseEventArgs e)
+        {
+            if(radioButton1.Checked)
+            {
+                if (e.RowIndex > -1)
+                {
+                    ViewDeb.Rows[e.RowIndex].Selected = true;
+                    getBusinessDebtor();
+                }
+                else
+                { }
+                if (radioButton2.Checked)
+                {
+                    if (e.RowIndex > -1)
+                    {
+                        ViewDeb.Rows[e.RowIndex].Selected = true;
+                        getPersonalDebtor();
+                    }
+                }
+                else
+                { }
+
             }
         }
     }
